Use a parameter to keep track of the template's content hash
Using tags to keep track of the content hash has the nice property of
allowing us to publish unmodified template files (kept in the __raw__
attribute), but has the very unpleasant effect that every tagged
resource is marked as modified when there's an update.

Parameter changes don't affect resources, avoiding unnecessary noise in
the change set, but we do need to inject the parameter in the template,
forcing us to change it.

For now, the latter option is taken to be the lesser evil. In the future
we can look at giving back control to the user by checking for a
user-defined parameter in the template.