Allow for stacks to be managed in independent projects
The metadata parameter associated with stacks now includes a project
identifier, allowing different stacks to be managed in independent
projects.

In particular, stacks that are not managed in the current project will
not be flagged as _orphaned_.

While we don't currently do anything about orphaned stacks, the intent
is that dropping a stack from a project flags it for deletion. As change
sets for stack deletion are not currently supported, it is left as an
exercise how this will be handled in practice.