Make it easier to deploy region-specific stacks
The AWS region can now be set/overridden directly in individual stacks.
Region information in deployment targets will function as a default, to
support this.

In the internal model, a target's 'region' now gives way to a
'default-region' setting, and targets are no longer expanded on region
alone. Instead, the 'stack' map now gains an extra 'region' level, and
tracks stacks across all regions for a single target.