Let's allow list concatenation
A deep merge of two lists could be interpreted in multiple ways. For
instance, it could reasonably:

- replace the first list with the second;
- produce a combined list by appending the elements of the second list
  to the first;
- deep merge corresponding elements in each list;
- (do something else?)

Without specialized syntax or operators, only one of the options can be
implemented. So far, the approach has been to disallow the construct
altogether, and throw an exception.

Of the options listed, I've mostly found a need for the first two:
replace, or concatenate.

The first alternative, to replace the original list, can be said to go
against the additive nature of the deep_merge operation. Particularly so
as it is used here to build templates from partials.

This alternative can also be achieved today by splitting templates into
finer-grained alternative lists, and picking only the partial template
with the relevant list.

To implement the second alternative, of concatenating lists, today, one
would have to separately maintain alternative lists of elements,
potentially duplicating portions of these lists. This goes against the
intent to build templates by composing partials.

So, we go with that.