Let _version.py print out the version information
In particular, this file depends only on the standard python library,
and no other modules in cfn_review_bot, allowing it to be used without
other dependencies installed.

A special 'github-action' mode is added that can be used in Github
Actions to make package_version and git_revision available, as output
variables, to subsequent actions.