feat: skip no-op changesets from markdown summary
No-op changesets to stack templates are now explicitly mentioned in
target summaries; details for these are omitted in the markdown summary
report. This reduces noise, and the need to perform manual sync
operations when CloudFormation refuses to forcefully update a stack
template.

As part of this change, the bot will now wait for changesets to be ready
whenever they're generated, and not just when the markdown summary is
requested.

Information about no-op changesets is only available in non `--dry-run`
mode as we rely on CloudFormation to determine whether an updated
template results in a no-op changeset.

This change allows for templates in a repository to silently diverge
from what's deployed by CloudFormation, as long as changes remain no-op
from CloudFormation's point-of-view. For instance, updates to mappings
that don't affect a particular account/region translate to no-op
changesets.